Remove client-side transformations on config input
