+++ server/lib/get-browsers.js
@@ -125,11 +125,50 @@ function importJSON(path) {
   return JSON.parse(readFileSync(new URL(path, import.meta.url)))
 }
 
-function configToQuery(query) {
-  return query
+function configToQuery(config) {
+  if (hasJSONSymbols(config)) {
+    return jsonConfigToQuery(config)
+  }
+
+  return rcConfigToQuery(config)
+}
+
+const JSON_FRAGMENT_REQUIRED_SYMBOLS = [':', '[', ']']
+
+function jsonConfigToQuery(config) {
+  try {
+    return JSONToQuery(config)
+  } catch {}
+  try {
+    return JSONFragmentToQuery(config)
+  } catch {}
+
+  return config
+}
+
+function rcConfigToQuery(config) {
+  return config
     .toString()
     .replace(/#[^\n]*/g, '')
     .split(/\n|,/)
     .map(line => line.trim())
     .filter(line => line !== '')
+}
+
+function JSONToQuery(jsonConfig) {
+  let data = JSON.parse(jsonConfig).browserslist
+  return data.join(',')
+}
+
+function JSONFragmentToQuery(jsonFragment) {
+  return JSONToQuery('{' + jsonFragment + '}')
+}
+
+function hasJSONSymbols(string) {
+  for (let symbol of JSON_FRAGMENT_REQUIRED_SYMBOLS) {
+    if (!string.includes(symbol)) {
+      return false
+    }
+  }
+  return true
 }
